requirements.txt.txt


pandas
plotly_express
streamlit